fix(oracle): series add/sub/mul/div infer operand dtype, don't force float64
op_series_binary_numeric hardcoded dtype="float64" on both operands, so the
live-oracle gate reported Float64 results even for genuine int outputs:
int+int over a full label overlap is int64 in real pandas (and in FP, and in
the stored fixture), but the oracle returned float64 -> a false live-gate red
for every int-valued series_add/sub/mul case.

Let pandas infer the operand dtype from the values, mirroring real
`pd.Series([...])` construction exactly: all-int operands stay int64, a
None-containing numeric operand coerces to float64+NaN (so alignment-
introduced misses still serialize as na_n), and division is always float.

Found by running the full corpus through the live differential gate
(--oracle live). Verified in-process baseline-vs-modified over 1252 fixtures:
0 regressions, +8 (series_add full-overlap/duplicate-label, series_mul
aligned). Replay/CI is unaffected (it uses stored fixtures, not the oracle).

Part of br-frankenpandas-zozby (oracle fidelity).

+++ crates/fp-conformance/oracle/pandas_oracle.py
@@ -404,8 +404,14 @@ def op_series_binary_numeric(
     left_values = [scalar_from_json(item) for item in left["values"]]
     right_values = [scalar_from_json(item) for item in right["values"]]
 
-    lhs = pd.Series(left_values, index=left_index, dtype="float64")
-    rhs = pd.Series(right_values, index=right_index, dtype="float64")
+    # Let pandas infer the operand dtype (do NOT force float64). This mirrors
+    # real `pd.Series([...])` list construction exactly: an all-int operand
+    # stays int64 (so int+int full-overlap yields int64, not float64), a
+    # None-containing numeric operand coerces to float64 with NaN, and division
+    # is always float. The old forced-float64 made the live oracle report
+    # Float64 even for genuine int results, diverging from real pandas and FP.
+    lhs = pd.Series(left_values, index=left_index)
+    rhs = pd.Series(right_values, index=right_index)
     if operation == "series_add":
         out = lhs + rhs
     elif operation == "series_sub":
